move elf creating to prevent circular imports
