+++ pwndbg/aglib/elf.py
@@ -9,8 +9,13 @@ working on a BSD system which simply does not have /proc.
 from __future__ import annotations
 
 import ctypes
+import subprocess
 import importlib
 import sys
+import tempfile
+import pwndbg.lib
+import pwndbg.lib.zig
+import pwndbg.lib.config
 from typing import Dict
 from typing import List
 from typing import NamedTuple
@@ -487,3 +492,77 @@ def map_inner(ei_class: int, ehdr: Ehdr, objfile: str) -> Tuple[pwndbg.lib.memor
         page.objfile = objfile
 
     return tuple(sorted(pages))
+
+
+gcc_compiler_path = pwndbg.config.add_param(
+    "gcc-compiler-path",
+    "",
+    "path to the gcc/g++ toolchain for generating imported symbols",
+    param_class=pwndbg.lib.config.PARAM_OPTIONAL_FILENAME,
+)
+
+def compile_with_flags(gcc_extra_flags):
+    if gcc_compiler_path != "":
+        compiler_flags = [gcc_compiler_path]
+    else:
+        try:
+            compiler_flags = pwndbg.lib.zig.flags(pwndbg.aglib.arch)
+        except ValueError as exception:
+            print(message.error(exception))
+            return False
+
+    gcc_cmd = compiler_flags + gcc_extra_flags
+
+    try:
+        subprocess.run(gcc_cmd, check=True, text=True)
+        return True
+    except subprocess.CalledProcessError as exception:
+        print(message.error(exception))
+        print(
+            message.error(
+                f"Failed to compile {gcc_extra_flags[0]}. Please fix any compilation errors there may be."
+            )
+        )
+    except Exception as exception:
+        print(message.error(exception))
+        print(message.error("An error occured while generating the debug symbols."))
+    return False
+
+
+def create_blank_elf():
+    try:
+        import lief
+    except ImportError:
+        print(
+            message.error(
+                "lief python package is not installed (this will be auto-installed with next version of Pwndbg; for now, you can install it manually in the Pwndbg venv)."
+            )
+        )
+        return None
+    tmp = tempfile.NamedTemporaryFile(delete=False, suffix=".S")
+    tmp.write(b".global _start\n_start:\nnop")
+    tmp.flush()
+    _, output_path = tempfile.mkstemp(prefix="blank-", suffix=".dbg")
+
+    gcc_extra_flags = [
+        tmp.name,
+        "-nostdlib",
+        "--static",
+        "-o",
+        output_path,
+    ]
+
+    if not compile_with_flags(gcc_extra_flags):
+        return None
+
+    blank_elf = lief.ELF.parse(output_path)
+    if blank_elf is None:
+        print(message.error("failed to parse blank elf"))
+        return
+
+    for s in blank_elf.symbols:
+        blank_elf.remove_symtab_symbol(s)
+
+    blank_elf.write(output_path)
+
+    return output_path
